refactor(cb2-0000): split tech record service
