+++ src/app/features/search/multiple-search-results/multiple-search-results.component.ts
@@ -3,10 +3,11 @@ import { Component, OnDestroy } from '@angular/core';
 import { GlobalErrorService } from '@core/components/global-error/global-error.service';
 import { VehicleTechRecordModel } from '@models/vehicle-tech-record.model';
 import { select, Store } from '@ngrx/store';
-import { SEARCH_TYPES, TechnicalRecordService } from '@services/technical-record/technical-record.service';
 import { selectQueryParams } from '@store/router/selectors/router.selectors';
 import { Observable, Subject, takeUntil } from 'rxjs';
 import { Roles } from '@models/roles.enum';
+import { SEARCH_TYPES, TechnicalRecordHttpService } from '@services/technical-record-http/technical-record-http.service';
+import { TechnicalRecordService } from '@services/technical-record/technical-record.service';
 
 @Component({
   selector: 'app-multiple-search-results',
@@ -20,6 +21,7 @@ export class MultipleSearchResultsComponent implements OnDestroy {
   constructor(
     public globalErrorService: GlobalErrorService,
     private technicalRecordService: TechnicalRecordService,
+    private technicalRecordHttpService: TechnicalRecordHttpService,
     private store: Store,
     private location: Location
   ) {
@@ -30,7 +32,7 @@ export class MultipleSearchResultsComponent implements OnDestroy {
 
         if (searchTerm && Object.values(SEARCH_TYPES).includes(type)) {
           this.globalErrorService.clearErrors();
-          this.technicalRecordService.searchBy(type, searchTerm);
+          this.technicalRecordHttpService.searchBy(type, searchTerm);
         }
       }
     });
