+++ src/app/features/tech-record/components/tech-record-amend-vin/tech-record-amend-vin.component.ts
@@ -45,7 +45,7 @@ export class AmendVinComponent implements OnDestroy {
         },
         '',
         [CustomValidators.alphanumeric(), Validators.minLength(3), Validators.maxLength(21), Validators.required],
-        [this.technicalRecordService.validateVin(this.vehicle?.vin)]
+        [this.technicalRecordService.validateVinForUpdate(this.vehicle?.vin)]
       )
     });
