+++ src/app/features/tech-record/create/create-tech-record.component.ts
@@ -9,7 +9,9 @@ import { CustomFormControl, CustomFormGroup, FormNodeTypes } from '@forms/servic
 import { CustomValidators } from '@forms/validators/custom-validators';
 import { StatusCodes, TechRecordModel, VehicleTechRecordModel, VehicleTypes } from '@models/vehicle-tech-record.model';
 import { Store } from '@ngrx/store';
-import { SEARCH_TYPES, TechnicalRecordService } from '@services/technical-record/technical-record.service';
+import { BatchTechnicalRecordService } from '@services/batch-technical-record/batch-technical-record.service';
+import { SEARCH_TYPES } from '@services/technical-record-http/technical-record-http.service';
+import { TechnicalRecordService } from '@services/technical-record/technical-record.service';
 import { setSpinnerState } from '@store/spinner/actions/spinner.actions';
 import { firstValueFrom } from 'rxjs';
 
@@ -68,11 +70,12 @@ export class CreateTechRecordComponent implements OnChanges {
   constructor(
     private globalErrorService: GlobalErrorService,
     private technicalRecordService: TechnicalRecordService,
+    private batchTechRecordService: BatchTechnicalRecordService,
     private route: ActivatedRoute,
     private router: Router,
     private store: Store
   ) {
-    this.technicalRecordService.clearBatch();
+    this.batchTechRecordService.clearBatch();
   }
 
   ngOnChanges(): void {
