+++ src/app/features/test-records/amend/amend-test-records-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { DefectSelectComponent } from '@forms/components/defect-select/defect-select.component';
 import { DefectComponent } from '@forms/custom-sections/defect/defect.component';
 import { CancelEditTestGuard } from '@guards/cancel-edit-test/cancel-edit-test.guard';
-import { RoleGuard } from '@guards/roles.guard';
+import { RoleGuard } from '@guards/role-guard/roles.guard';
 import { Roles } from '@models/roles.enum';
 import { DefectsTaxonomyResolver } from 'src/app/resolvers/defects-taxonomy/defects-taxonomy.resolver';
 import { TestResultResolver } from 'src/app/resolvers/test-result/test-result.resolver';
