+++ src/app/services/technical-record/technical-record.service.spec.ts
@@ -1,17 +1,14 @@
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
-import { fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { FormControl, FormGroup, ValidationErrors } from '@angular/forms';
+import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { CustomFormControl, FormNodeTypes } from '@forms/services/dynamic-form.types';
 import { mockVehicleTechnicalRecord, mockVehicleTechnicalRecordList } from '@mocks/mock-vehicle-technical-record.mock';
 import { StatusCodes, VehicleTechRecordModel, VehicleTypes } from '@models/vehicle-tech-record.model';
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
+import { SEARCH_TYPES } from '@services/technical-record-http/technical-record-http.service';
 import { initialAppState, State } from '@store/index';
 import { editableVehicleTechRecord, selectVehicleTechnicalRecordsBySystemNumber, updateEditingTechRecord } from '@store/technical-records';
-import cloneDeep from 'lodash.clonedeep';
-import { firstValueFrom, lastValueFrom, Observable, of } from 'rxjs';
 import { environment } from '../../../environments/environment';
-import { SEARCH_TYPES, TechnicalRecordService } from './technical-record.service';
+import { TechnicalRecordService } from './technical-record.service';
 
 describe('TechnicalRecordService', () => {
   let service: TechnicalRecordService;
@@ -35,257 +32,6 @@ describe('TechnicalRecordService', () => {
 
   it('should be created', () => {
     expect(service).toBeTruthy();
-  });
-
-  describe('API', () => {
-    describe('getByVin', () => {
-      it('should get an array of matching results', () => {
-        const searchParams = { searchTerm: 'A_VIN', type: 'vin' };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.getByVin(searchParams.searchTerm).subscribe(response => {
-          expect(response).toEqual(mockData);
-        });
-
-        // Check for correct requests: should have made one request to search from expected URL
-        const req = httpClient.expectOne(
-          `${environment.VTM_API_URI}/vehicles/${searchParams.searchTerm}/tech-records?status=all&metadata=true&searchCriteria=vin`
-        );
-        expect(req.request.method).toEqual('GET');
-
-        // Provide each request with a mock response
-        req.flush(mockData);
-      });
-
-      it('should handle errors', () => {
-        const searchParams = { searchTerm: 'A_VIN', type: 'vin' };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.getByVin(searchParams.searchTerm).subscribe(response => {
-          expect(response).toEqual(mockData);
-        });
-
-        // Check for correct requests: should have made one request to search from expected URL
-        const req = httpClient.expectOne(
-          `${environment.VTM_API_URI}/vehicles/${searchParams.searchTerm}/tech-records?status=all&metadata=true&searchCriteria=vin`
-        );
-        expect(req.request.method).toEqual('GET');
-
-        // Respond with mock error
-        req.flush('Deliberate 500 error', { status: 500, statusText: 'Server Error' });
-      });
-    });
-
-    describe('getByPartialVin', () => {
-      it('should get an array of matching results', () => {
-        const searchParams = { searchTerm: 'A_VIN', type: 'partialVin' };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.getByPartialVin(searchParams.searchTerm).subscribe(response => {
-          expect(response).toEqual(mockData);
-        });
-
-        // Check for correct requests: should have made one request to search from expected URL
-        const req = httpClient.expectOne(
-          `${environment.VTM_API_URI}/vehicles/${searchParams.searchTerm}/tech-records?status=all&metadata=true&searchCriteria=partialVin`
-        );
-        expect(req.request.method).toEqual('GET');
-
-        // Provide each request with a mock response
-        req.flush(mockData);
-      });
-
-      it('should handle errors', () => {
-        const searchParams = { searchTerm: 'A_VIN', type: 'partialVin' };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.getByPartialVin(searchParams.searchTerm).subscribe(response => {
-          expect(response).toEqual(mockData);
-        });
-
-        // Check for correct requests: should have made one request to search from expected URL
-        const req = httpClient.expectOne(
-          `${environment.VTM_API_URI}/vehicles/${searchParams.searchTerm}/tech-records?status=all&metadata=true&searchCriteria=partialVin`
-        );
-        expect(req.request.method).toEqual('GET');
-
-        // Respond with mock error
-        req.flush('Deliberate 500 error', { status: 500, statusText: 'Server Error' });
-      });
-    });
-
-    describe('getByVrm', () => {
-      it('should get an array of matching results', () => {
-        const params = { term: 'A_VRM', type: 'vrm' };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.getByVrm(params.term).subscribe(response => expect(response).toEqual(mockData));
-
-        // Check for correct requests: should have made one request to search from expected URL
-        const req = httpClient.expectOne(
-          `${environment.VTM_API_URI}/vehicles/${params.term}/tech-records?status=all&metadata=true&searchCriteria=${params.type}`
-        );
-        expect(req.request.method).toEqual('GET');
-
-        // Provide each request with a mock response
-        req.flush(mockData);
-      });
-
-      it('should handle errors', () => {
-        const params = { term: 'A_VRM', type: 'vrm' };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.getByVrm(params.term).subscribe(response => expect(response).toEqual(mockData));
-
-        // Check for correct requests: should have made one request to search from expected URL
-        const req = httpClient.expectOne(
-          `${environment.VTM_API_URI}/vehicles/${params.term}/tech-records?status=all&metadata=true&searchCriteria=${params.type}`
-        );
-        expect(req.request.method).toEqual('GET');
-
-        // Respond with mock error
-        req.flush('Deliberate 500 error', { status: 500, statusText: 'Server Error' });
-      });
-    });
-
-    describe('getByTrailerId', () => {
-      it('should get an array of matching results', () => {
-        const params = { term: 'A_TRAILER_ID', type: 'trailerId' };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.getByTrailerId(params.term).subscribe(response => expect(response).toEqual(mockData));
-
-        // Check for correct requests: should have made one request to search from expected URL
-        const req = httpClient.expectOne(
-          `${environment.VTM_API_URI}/vehicles/${params.term}/tech-records?status=all&metadata=true&searchCriteria=${params.type}`
-        );
-        expect(req.request.method).toEqual('GET');
-
-        // Provide each request with a mock response
-        req.flush(mockData);
-      });
-
-      it('should handle errors', () => {
-        const params = { term: 'A_TRAILER_ID', type: 'trailerId' };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.getByTrailerId(params.term).subscribe(response => expect(response).toEqual(mockData));
-
-        // Check for correct requests: should have made one request to search from expected URL
-        const req = httpClient.expectOne(
-          `${environment.VTM_API_URI}/vehicles/${params.term}/tech-records?status=all&metadata=true&searchCriteria=${params.type}`
-        );
-        expect(req.request.method).toEqual('GET');
-
-        // Respond with mock error
-        req.flush('Deliberate 500 error', { status: 500, statusText: 'Server Error' });
-      });
-    });
-
-    describe('createVehicleRecord', () => {
-      it('should call post with the correct URL, body and response type', fakeAsync(() => {
-        const expectedVehicle = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1)[0];
-        const expectedUser = { name: 'test', id: '1234' };
-
-        service.createVehicleRecord(expectedVehicle, expectedUser).subscribe();
-
-        const expectedBody = {
-          msUserDetails: { msOid: expectedUser.id, msUser: expectedUser.name },
-          vin: expectedVehicle.vin,
-          primaryVrm: expectedVehicle.vrms ? expectedVehicle.vrms[0].vrm : null,
-          trailerId: expectedVehicle.trailerId ?? null,
-          techRecord: expectedVehicle.techRecord
-        };
-
-        const request = httpClient.expectOne(`${environment.VTM_API_URI}/vehicles`);
-
-        expect(request.request.method).toEqual('POST');
-        expect(request.request.body).toEqual(expectedBody);
-
-        request.flush(expectedVehicle);
-      }));
-
-      it('should return an array with the newly created vehicle record', () => {
-        const expectedVehicle = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1)[0];
-
-        const expectedResult = {
-          vin: expectedVehicle.vin,
-          primaryVrm: expectedVehicle.vrms ? expectedVehicle.vrms[0].vrm : null,
-          trailerId: expectedVehicle.trailerId ?? null,
-          techRecord: expectedVehicle.techRecord
-        };
-
-        expect(lastValueFrom(service.createVehicleRecord(expectedVehicle, { name: 'test', id: '1234' }))).resolves.toEqual(expectedResult);
-
-        const request = httpClient.expectOne(`${environment.VTM_API_URI}/vehicles`);
-        request.flush(expectedResult);
-      });
-    });
-
-    describe('createProvisionalTechRecord', () => {
-      it('should return an array with a new tech record having added provisional', fakeAsync(() => {
-        const params = { systemNumber: '12345', user: { name: 'TEST', id: '1234' } };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.createProvisionalTechRecord(params.systemNumber, mockData[0].techRecord[0], params.user).subscribe();
-
-        // Check for correct requests: should have made one request to the PUT URL
-        const req = httpClient.expectOne(`${environment.VTM_API_URI}/vehicles/add-provisional/${params.systemNumber}`);
-        expect(req.request.method).toEqual('POST');
-
-        // Provide each request with a mock response
-        req.flush(mockData);
-      }));
-    });
-
-    describe('updateTechRecords', () => {
-      it('should return an array with a new tech record and updated status code', fakeAsync(() => {
-        const params = { systemNumber: '12345', user: { name: 'TEST', id: '1234' }, oldStatusCode: StatusCodes.PROVISIONAL };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.updateTechRecords(params.systemNumber, mockData[0], params.user, params.oldStatusCode).subscribe();
-
-        // Check for correct requests: should have made one request to the PUT URL
-        const req = httpClient.expectOne(`${environment.VTM_API_URI}/vehicles/${params.systemNumber}?oldStatusCode=${params.oldStatusCode}`);
-        expect(req.request.method).toEqual('PUT');
-
-        // should format the vrms for the update payload
-        expect(req.request.body).toHaveProperty('primaryVrm');
-        expect(req.request.body).toHaveProperty('secondaryVrms');
-        expect(req.request.body).not.toHaveProperty('vrms');
-
-        // Provide each request with a mock response
-        req.flush(mockData);
-      }));
-
-      it('should return an array with a new tech record and updated status code using basic URL', fakeAsync(() => {
-        const params = { systemNumber: '12345', user: { name: 'TEST', id: '1234' } };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.updateTechRecords(params.systemNumber, mockData[0], params.user).subscribe();
-
-        // Check for correct requests: should have made one request to the PUT URL
-        const req = httpClient.expectOne(`${environment.VTM_API_URI}/vehicles/${params.systemNumber}`);
-        expect(req.request.method).toEqual('PUT');
-
-        // Provide each request with a mock response
-        req.flush(mockData);
-      }));
-    });
-
-    describe('updateVin', () => {
-      it('should make POST request to correct URL', fakeAsync(() => {
-        const params = { systemNumber: '12345', vin: 'MYNEWVIN', user: { name: 'TEST', id: '1234' } };
-        service.updateVin(params.vin, params.systemNumber, params.user).subscribe();
-
-        const req = httpClient.expectOne(`${environment.VTM_API_URI}/vehicles/update-vin/${params.systemNumber}`);
-        expect(req.request.method).toEqual('PUT');
-      }));
-    });
-
-    describe('archiveTechRecord', () => {
-      it('should return a new tech record having added provisional', fakeAsync(() => {
-        const params = { systemNumber: '12345', reasonForArchiving: 'some reason', user: { name: 'TEST', id: '1234' } };
-        const mockData = mockVehicleTechnicalRecordList(VehicleTypes.PSV, 1);
-        service.archiveTechnicalRecord(params.systemNumber, mockData[0].techRecord[0], params.reasonForArchiving, params.user).subscribe();
-
-        // Check for correct requests: should have made one request to the PUT URL
-        const req = httpClient.expectOne(`${environment.VTM_API_URI}/vehicles/archive/${params.systemNumber}`);
-        expect(req.request.method).toEqual('PUT');
-
-        // Provide each request with a mock response
-        req.flush(mockData);
-      }));
-    });
   });
 
   describe('isUnique', () => {
@@ -381,15 +127,6 @@ describe('TechnicalRecordService', () => {
     });
   });
 
-  describe('calculateTechRecordToUpdate', () => {
-    let vehicle: VehicleTechRecordModel = {
-      vin: '123',
-      vrms: [{ vrm: '123', isPrimary: true }],
-      systemNumber: '1234',
-      techRecord: []
-    };
-  });
-
   describe('business logic methods', () => {
     describe('updateEditingTechRecord', () => {
       it('should patch the missing information for the technical record and dispatch the action to update the editing vehicle record with the full vehicle record', () => {
@@ -440,139 +177,6 @@ describe('TechnicalRecordService', () => {
         service.updateEditingTechRecord(mockVehicleRecord);
         expect(dispatchSpy).toHaveBeenCalledTimes(1);
         expect(dispatchSpy).toHaveBeenCalledWith(updateEditingTechRecord({ vehicleTechRecord: mockVehicleRecord }));
-      });
-    });
-  });
-
-  describe('validateForBatch', () => {
-    let testGroup: FormGroup;
-    beforeEach(() => {
-      testGroup = new FormGroup({
-        vin: new CustomFormControl({ name: 'vin', type: FormNodeTypes.CONTROL }, null, null),
-        trailerId: new FormControl({ name: 'trailerId', value: '' }, null),
-        systemNumber: new FormControl({ name: 'systemNumber', value: '' }, null)
-      });
-    });
-
-    it('return null if vin and trailer id are not provided', done => {
-      expect.assertions(1);
-      testGroup.get('vin')!.setValue(null);
-      testGroup.get('trailerId')!.setValue(null);
-      const serviceCall = service.validateForBatch()(testGroup.get('vin')!);
-      (serviceCall as Observable<ValidationErrors | null>).subscribe(errors => {
-        expect(errors).toEqual(null);
-        done();
-      });
-    });
-
-    it('return null if the required controls do not exist', done => {
-      expect.assertions(1);
-      testGroup = new FormGroup({ vin: new FormControl({ name: 'vin' }) });
-      const serviceCall = service.validateForBatch()(testGroup.get('vin')!);
-      (serviceCall as Observable<ValidationErrors | null>).subscribe(errors => {
-        expect(errors).toEqual(null);
-        done();
-      });
-    });
-
-    it('throws an error if trailer id is provided but vin is not', done => {
-      expect.assertions(1);
-      testGroup.get('trailerId')!.setValue('test');
-
-      const serviceCall = service.validateForBatch()(testGroup.get('vin')!);
-      (serviceCall as Observable<ValidationErrors | null>).subscribe(errors => {
-        expect(errors).toEqual({ validateForBatch: { message: 'VIN is required' } });
-        done();
-      });
-    });
-
-    describe('when only vin is provided', () => {
-      it('should return null when it is a unique vin', async () => {
-        testGroup.get('trailerId')!.setValue('');
-        testGroup.get('vin')!.setValue('TESTVIN');
-
-        const isUniqueSpy = jest.spyOn(service, 'isUnique').mockReturnValueOnce(of(true));
-
-        const serviceCall = service.validateForBatch()(testGroup.get('vin')!) as Observable<ValidationErrors | null>;
-        const errors = await firstValueFrom(serviceCall);
-        expect(isUniqueSpy).toHaveBeenCalled();
-        expect(errors).toBeNull();
-      });
-
-      it('should set a warning when it is not a unique vin', async () => {
-        testGroup.get('trailerId')!.setValue('');
-        const vinControl = testGroup.get('vin') as CustomFormControl;
-        vinControl.setValue('TESTVIN');
-        jest.spyOn(service, 'isUnique').mockReturnValueOnce(of(false));
-        await firstValueFrom(service.validateForBatch()(testGroup.get('vin')!) as Observable<ValidationErrors | null>);
-        expect(vinControl.meta.warning).toEqual('This VIN already exists, if you continue it will be associated with two vehicles');
-      });
-    });
-
-    describe('when both vin and trailer id are provided', () => {
-      it('returns null if only 1 vehicle exists with those values with no current tech record', async () => {
-        expect.assertions(1);
-        testGroup.get('vin')!.setValue('TESTVIN');
-        testGroup.get('trailerId')!.setValue('TESTTRAILERID');
-        const mockVehicleRecord = {
-          vin: 'TESTVIN',
-          trailerId: 'TESTTRAILERID',
-          systemNumber: 'TESTSYSTEMNUMBER',
-          techRecord: []
-        } as unknown as VehicleTechRecordModel;
-
-        jest.spyOn(service, 'getByVin').mockReturnValue(of([mockVehicleRecord]));
-
-        const serviceCall = service.validateForBatch()(testGroup.get('vin')!) as Observable<ValidationErrors | null>;
-        const errors = await firstValueFrom(serviceCall);
-        expect(errors).toBeNull();
-      });
-
-      it('returns null if only 1 vehicle exists with those values with a current tech record', async () => {
-        expect.assertions(1);
-        testGroup.get('vin')!.setValue('TESTVIN');
-        testGroup.get('trailerId')!.setValue('TESTTRAILERID');
-        const mockVehicleRecord = {
-          vin: 'TESTVIN',
-          trailerId: 'TESTTRAILERID',
-          systemNumber: 'TESTSYSTEMNUMBER',
-          techRecord: [{ statusCode: StatusCodes.CURRENT }]
-        } as unknown as VehicleTechRecordModel;
-
-        jest.spyOn(service, 'getByVin').mockReturnValue(of([mockVehicleRecord]));
-
-        const serviceCall = service.validateForBatch()(testGroup.get('vin')!) as Observable<ValidationErrors | null>;
-        const errors = await firstValueFrom(serviceCall);
-        expect(errors).toEqual({ validateForBatch: { message: 'This record cannot be updated as it has a Current tech record' } });
-      });
-
-      it('throws an error if more than 1 vehicle exists with those values', done => {
-        expect.assertions(1);
-        testGroup.get('vin')!.setValue('TESTVIN');
-        testGroup.get('trailerId')!.setValue('TESTTRAILERID');
-        const mockVehicleRecord = { vin: 'TESTVIN', trailerId: 'TESTTRAILERID' } as VehicleTechRecordModel;
-
-        jest.spyOn(service, 'getByVin').mockReturnValue(of([cloneDeep(mockVehicleRecord), cloneDeep(mockVehicleRecord)]));
-
-        const serviceCall = service.validateForBatch()(testGroup.get('vin')!);
-        (serviceCall as Observable<ValidationErrors | null>).subscribe(errors => {
-          expect(errors).toEqual({ validateForBatch: { message: 'More than one vehicle has this VIN and Trailer ID' } });
-          done();
-        });
-      });
-
-      it('throws an error if no vehicle exists with those values', done => {
-        expect.assertions(1);
-        testGroup.get('vin')!.setValue('TESTVIN');
-        testGroup.get('trailerId')!.setValue('TESTTRAILERID');
-
-        jest.spyOn(service, 'getByVin').mockReturnValue(of([]));
-
-        const serviceCall = service.validateForBatch()(testGroup.get('vin')!);
-        (serviceCall as Observable<ValidationErrors | null>).subscribe(errors => {
-          expect(errors).toEqual({ validateForBatch: { message: 'Could not find a record with matching VIN and Trailer ID' } });
-          done();
-        });
       });
     });
   });
