+++ src/app/store/router/selectors/router.selectors.ts
@@ -19,6 +19,8 @@ export const {
 export const routerState = createSelector(selectRouter, state => state);
 export const currentRouteState = createSelector(selectCurrentRoute, state => state);
 
+export const selectRouteDataProperty = (property: string) => createSelector(selectRouteData, data => data[property]);
+
 export const selectRouteNestedParams = createSelector(selectRouter, router => {
   let currentRoute = router?.state?.root;
   let params: Params = {};
