+++ src/app/store/technical-records/effects/technical-record-service.effects.ts
@@ -12,6 +12,8 @@ import {
 } from '@models/vehicle-tech-record.model';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { select, Store } from '@ngrx/store';
+import { BatchTechnicalRecordService } from '@services/batch-technical-record/batch-technical-record.service';
+import { TechnicalRecordHttpService } from '@services/technical-record-http/technical-record-http.service';
 import { TechnicalRecordService } from '@services/technical-record/technical-record.service';
 import { UserService } from '@services/user-service/user-service';
 import { State } from '@store/index';
@@ -66,7 +68,9 @@ import { editableTechRecord, selectVehicleTechnicalRecordsBySystemNumber } from 
 export class TechnicalRecordServiceEffects {
   constructor(
     private actions$: Actions,
+    private techRecordHttpService: TechnicalRecordHttpService,
     private technicalRecordService: TechnicalRecordService,
+    private batchTechRecordService: BatchTechnicalRecordService,
     private userService: UserService,
     private store: Store<State>,
     private dfs: DynamicFormService
@@ -80,31 +84,31 @@ export class TechnicalRecordServiceEffects {
 
         switch (action.type) {
           case getByVin.type:
-            return this.technicalRecordService.getByVin(action.vin).pipe(
+            return this.techRecordHttpService.getByVin(action.vin).pipe(
               map(vehicleTechRecords => getByVinSuccess({ vehicleTechRecords })),
               catchError(error => of(getByVinFailure({ error: this.getTechRecordErrorMessage(error, 'getTechnicalRecords', 'vin'), anchorLink })))
             );
           case getByPartialVin.type:
-            return this.technicalRecordService.getByPartialVin(action.partialVin).pipe(
+            return this.techRecordHttpService.getByPartialVin(action.partialVin).pipe(
               map(vehicleTechRecords => getByPartialVinSuccess({ vehicleTechRecords })),
               catchError(error =>
                 of(getByPartialVinFailure({ error: this.getTechRecordErrorMessage(error, 'getTechnicalRecords', 'partialVin'), anchorLink }))
               )
             );
           case getByVrm.type:
-            return this.technicalRecordService.getByVrm(action.vrm).pipe(
+            return this.techRecordHttpService.getByVrm(action.vrm).pipe(
               map(vehicleTechRecords => getByVrmSuccess({ vehicleTechRecords })),
               catchError(error => of(getByVrmFailure({ error: this.getTechRecordErrorMessage(error, 'getTechnicalRecords', 'vrm'), anchorLink })))
             );
           case getByTrailerId.type:
-            return this.technicalRecordService.getByTrailerId(action.trailerId).pipe(
+            return this.techRecordHttpService.getByTrailerId(action.trailerId).pipe(
               map(vehicleTechRecords => getByTrailerIdSuccess({ vehicleTechRecords })),
               catchError(error =>
                 of(getByTrailerIdFailure({ error: this.getTechRecordErrorMessage(error, 'getTechnicalRecords', 'trailerId'), anchorLink }))
               )
             );
           case getBySystemNumber.type:
-            return this.technicalRecordService.getBySystemNumber(action.systemNumber).pipe(
+            return this.techRecordHttpService.getBySystemNumber(action.systemNumber).pipe(
               map(vehicleTechRecords => {
                 return getBySystemNumberSuccess({ vehicleTechRecords: vehicleTechRecords });
               }),
@@ -113,7 +117,7 @@ export class TechnicalRecordServiceEffects {
               )
             );
           case getByAll.type:
-            return this.technicalRecordService.getByAll(action.all).pipe(
+            return this.techRecordHttpService.getByAll(action.all).pipe(
               map(vehicleTechRecords => getByAllSuccess({ vehicleTechRecords })),
               catchError(error =>
                 of(
@@ -129,11 +133,11 @@ export class TechnicalRecordServiceEffects {
   createVehicleRecord$ = createEffect(() =>
     this.actions$.pipe(
       ofType(createVehicleRecord),
-      withLatestFrom(this.technicalRecordService.applicationId$, this.userService.name$, this.userService.id$),
+      withLatestFrom(this.batchTechRecordService.applicationId$, this.userService.name$, this.userService.id$),
       concatMap(([{ vehicle }, applicationId, name, id]) => {
         const vehicleRecord = { ...vehicle, techRecord: [{ ...vehicle.techRecord[0], applicationId }] };
 
-        return this.technicalRecordService.createVehicleRecord(vehicleRecord, { id, name }).pipe(
+        return this.techRecordHttpService.createVehicleRecord(vehicleRecord, { id, name }).pipe(
           map(response => createVehicleRecordSuccess({ vehicleTechRecords: [this.mapVehicleFromResponse(response)] })),
           catchError(error =>
             of(
@@ -154,7 +158,7 @@ export class TechnicalRecordServiceEffects {
       ofType(createProvisionalTechRecord),
       withLatestFrom(this.technicalRecordService.editableTechRecord$, this.userService.name$, this.userService.id$),
       switchMap(([action, record, name, id]) =>
-        this.technicalRecordService.createProvisionalTechRecord(action.systemNumber, record!, { id, name }).pipe(
+        this.techRecordHttpService.createProvisionalTechRecord(action.systemNumber, record!, { id, name }).pipe(
           map(vehicleTechRecord => createProvisionalTechRecordSuccess({ vehicleTechRecords: [vehicleTechRecord] })),
           catchError(error => of(createProvisionalTechRecordFailure({ error: this.getTechRecordErrorMessage(error, 'createProvisionalTechRecord') })))
         )
@@ -167,12 +171,10 @@ export class TechnicalRecordServiceEffects {
       ofType(updateTechRecords),
       withLatestFrom(this.technicalRecordService.editableVehicleTechRecord$, this.userService.name$, this.userService.id$),
       concatMap(([action, record, name, id]) =>
-        this.technicalRecordService
-          .updateTechRecords(action.systemNumber, record!, { id, name }, action.recordToArchiveStatus, action.newStatus)
-          .pipe(
-            map(vehicleTechRecord => updateTechRecordsSuccess({ vehicleTechRecords: [vehicleTechRecord] })),
-            catchError(error => of(updateTechRecordsFailure({ error: this.getTechRecordErrorMessage(error, 'updateTechnicalRecord') })))
-          )
+        this.techRecordHttpService.updateTechRecords(action.systemNumber, record!, { id, name }, action.recordToArchiveStatus, action.newStatus).pipe(
+          map(vehicleTechRecord => updateTechRecordsSuccess({ vehicleTechRecords: [vehicleTechRecord] })),
+          catchError(error => of(updateTechRecordsFailure({ error: this.getTechRecordErrorMessage(error, 'updateTechnicalRecord') })))
+        )
       )
     )
   );
@@ -182,7 +184,7 @@ export class TechnicalRecordServiceEffects {
       ofType(archiveTechRecord),
       withLatestFrom(this.technicalRecordService.editableTechRecord$, this.userService.name$, this.userService.id$),
       switchMap(([action, record, name, id]) =>
-        this.technicalRecordService.archiveTechnicalRecord(action.systemNumber, record!, action.reasonForArchiving, { id, name }).pipe(
+        this.techRecordHttpService.archiveTechnicalRecord(action.systemNumber, record!, action.reasonForArchiving, { id, name }).pipe(
           map(vehicleTechRecord => archiveTechRecordSuccess({ vehicleTechRecords: [vehicleTechRecord] })),
           catchError(error => of(archiveTechRecordFailure({ error: this.getTechRecordErrorMessage(error, 'archiveTechRecord') })))
         )
@@ -228,7 +230,7 @@ export class TechnicalRecordServiceEffects {
         this.userService.userEmail$
       ),
       switchMap(([{ reason }, vehicle, techRecord, name, id, email]) =>
-        this.technicalRecordService.generatePlate(vehicle!, techRecord!, reason, { name, id, email }).pipe(
+        this.techRecordHttpService.generatePlate(vehicle!, techRecord!, reason, { name, id, email }).pipe(
           map(() => generatePlateSuccess()),
           catchError(error => of(generatePlateFailure({ error: this.getTechRecordErrorMessage(error, 'generatePlate') })))
         )
@@ -247,7 +249,7 @@ export class TechnicalRecordServiceEffects {
         this.userService.userEmail$
       ),
       switchMap(([{ letterType, paragraphId }, vehicle, techRecord, name, id, email]) =>
-        this.technicalRecordService.generateLetter(vehicle!, techRecord!, letterType, paragraphId, { name, id, email }).pipe(
+        this.techRecordHttpService.generateLetter(vehicle!, techRecord!, letterType, paragraphId, { name, id, email }).pipe(
           map(value => generateLetterSuccess()),
           catchError(error => of(generateLetterFailure({ error: this.getTechRecordErrorMessage(error, 'generateLetter') })))
         )
@@ -260,7 +262,7 @@ export class TechnicalRecordServiceEffects {
       ofType(updateVin),
       withLatestFrom(this.userService.name$, this.userService.id$),
       switchMap(([{ newVin, systemNumber }, name, id]) =>
-        this.technicalRecordService.updateVin(newVin, systemNumber, { id, name }).pipe(
+        this.techRecordHttpService.updateVin(newVin, systemNumber, { id, name }).pipe(
           map(() => updateVinSuccess()),
           catchError(error => of(updateVinFailure({ error: error })))
         )
