+++ src/app/store/technical-records/selectors/technical-record-service.selectors.ts
@@ -1,4 +1,3 @@
-import { VehicleTechRecordModel } from '@models/vehicle-tech-record.model';
 import { createSelector } from '@ngrx/store';
 import { selectRouteNestedParams } from '@store/router/selectors/router.selectors';
 import { getVehicleTechRecordState } from '../reducers/technical-record-service.reducer';
@@ -21,8 +20,6 @@ export const selectVehicleTechnicalRecordsBySystemNumber = createSelector(
       return new Date(b).getTime() - new Date(a).getTime();
     };
 
-    return foundRecord
-      ? ({ ...foundRecord, techRecord: [...foundRecord.techRecord].sort((a, b) => sortByDate(a.createdAt, b.createdAt)) } as VehicleTechRecordModel)
-      : undefined;
+    return foundRecord ? { ...foundRecord, techRecord: [...foundRecord.techRecord].sort((a, b) => sortByDate(a.createdAt, b.createdAt)) } : undefined;
   }
 );
